+++ packages/kbn-i18n/src/react/provider.tsx
@@ -22,47 +22,7 @@ import * as React from 'react';
 import { IntlProvider } from 'react-intl';
 
 import * as i18n from '../core';
-import { isPseudoLocale, translateUsingPseudoLocale } from '../core/pseudo_locale';
-import { injectI18n } from './inject';
-
-/**
- * To translate label that includes nested `FormattedMessage` instances React Intl
- * replaces them with special placeholders (@__uid__@ELEMENT-uid-counter@__uid__@)
- * and maps them back with nested translations after `formatMessage` processes
- * original string, so we shouldn't modify these special placeholders with pseudo
- * translations otherwise React Intl won't be able to properly replace placeholders.
- * It's implementation detail of the React Intl, but since pseudo localization is dev
- * only feature we should be fine here.
- * @param message
- */
-function translateFormattedMessageUsingPseudoLocale(message: string) {
-  const formattedMessageDelimiter = message.match(/@__.{10}__@/);
-  if (formattedMessageDelimiter !== null) {
-    return message
-      .split(formattedMessageDelimiter[0])
-      .map(part => (part.startsWith('ELEMENT-') ? part : translateUsingPseudoLocale(part)))
-      .join(formattedMessageDelimiter[0]);
-  }
-
-  return translateUsingPseudoLocale(message);
-}
-
-/**
- * If pseudo locale is detected, default intl.formatMessage should be decorated
- * with the pseudo localization function.
- * @param child I18nProvider child component.
- */
-function wrapIntlFormatMessage(child: React.ReactElement) {
-  return React.createElement(
-    injectI18n(({ intl }) => {
-      const formatMessage = intl.formatMessage;
-      intl.formatMessage = (...args) =>
-        translateFormattedMessageUsingPseudoLocale(formatMessage(...args));
-
-      return React.Children.only(child);
-    })
-  );
-}
+import { PseudoLocaleWrapper } from './pseudo_locale_wrapper';
 
 /**
  * The library uses the provider pattern to scope an i18n context to a tree
@@ -81,9 +41,7 @@ export class I18nProvider extends React.PureComponent {
         formats={i18n.getFormats()}
         textComponent={React.Fragment}
       >
-        {isPseudoLocale(i18n.getLocale()) && React.isValidElement(this.props.children)
-          ? wrapIntlFormatMessage(this.props.children)
-          : this.props.children}
+        <PseudoLocaleWrapper>{this.props.children}</PseudoLocaleWrapper>
       </IntlProvider>
     );
   }
