+++ x-pack/legacy/plugins/siem/public/components/open_timeline/open_timeline_modal/index.test.tsx
@@ -22,17 +22,7 @@ const getStateChildComponent = (
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   wrapper: ReactWrapper<any, Readonly<{}>, React.Component<{}, {}, any>>
 ): // eslint-disable-next-line @typescript-eslint/no-explicit-any
-React.Component<{}, {}, any> =>
-  wrapper
-    .childAt(0)
-    .childAt(0)
-    .childAt(0)
-    .childAt(0)
-    .childAt(0)
-    .childAt(0)
-    .childAt(0)
-    .childAt(0)
-    .instance();
+React.Component<{}, {}, any> => wrapper.find('[data-test-subj="state-child-component"]').instance();
 
 describe('OpenTimelineModalButton', () => {
   const theme = () => ({ eui: euiDarkVars, darkMode: true });
@@ -64,7 +54,10 @@ describe('OpenTimelineModalButton', () => {
         <ThemeProvider theme={theme}>
           <TestProviderWithoutDragAndDrop>
             <MockedProvider mocks={mockOpenTimelineQueryResults} addTypename={false}>
-              <OpenTimelineModalButton onToggle={jest.fn()} />
+              <OpenTimelineModalButton
+                data-test-subj="state-child-component"
+                onToggle={jest.fn()}
+              />
             </MockedProvider>
           </TestProviderWithoutDragAndDrop>
         </ThemeProvider>
@@ -156,7 +149,10 @@ describe('OpenTimelineModalButton', () => {
         <ThemeProvider theme={theme}>
           <TestProviderWithoutDragAndDrop>
             <MockedProvider mocks={mockOpenTimelineQueryResults} addTypename={false}>
-              <OpenTimelineModalButton onToggle={jest.fn()} />
+              <OpenTimelineModalButton
+                data-test-subj="state-child-component"
+                onToggle={jest.fn()}
+              />
             </MockedProvider>
           </TestProviderWithoutDragAndDrop>
         </ThemeProvider>
